+++ addon/addon.setup.php
@@ -1,0 +1,12 @@
+<?php
+
+return [
+    'author' => 'Packet Tide, LLC',
+    'author_url' => 'https://packettide.com/',
+    'name' => 'Coilpack',
+    'description' => 'Supercharge ExpressionEngine with more power, features, and flexibility!',
+    'version' => \Composer\InstalledVersions::getPrettyVersion('expressionengine/coilpack'),
+    'namespace' => 'ExpressionEngine\Addons\Coilpack',
+    'settings_exist' => false,
+    'built_in' => true,
+];